Refactor Tests
Instead of requiring every class to implement
Create/CreateNew, use a single abstract method to
create an IPersistantQueueFactory.

Move tests into more logical units.

Rename tests again, because I apparently have nothing
better to do.
